update file here

+++ api/delete_bale_from_dispatch_note.php
@@ -25,45 +25,51 @@ $canDelete=0;
 
 
 
-    // $sql = "Select id from dispatch_note  where id=$dispatch_noteid and  open_close=0 and receicer_close=0";
-    // $result = $conn->query($sql);
+    $sql = "Select id from dispatch_note_total_received  where dispatch_noteid=$dispatch_noteid and  quantity=0 ";
+    $result = $conn->query($sql);
      
-    //  if ($result->num_rows > 0) {
-    //    // output data of each row
-    //    while($row = $result->fetch_assoc()) {
-    //     // echo "id: " . $row["id"]. " - Name: " . $row["firstname"]. " " . $row["lastname"]. "<br>";
-    //       $canDelete=$row["id"];
+     if ($result->num_rows > 0) {
+       // output data of each row
+       while($row = $result->fetch_assoc()) {
+        // echo "id: " . $row["id"]. " - Name: " . $row["firstname"]. " " . $row["lastname"]. "<br>";
+          $canDelete=$row["id"];
         
-    //    }
+       }
 
-    //  }
-
-
-
-    //   if ($canDelete==0) {
-        
-    //   }
+     }
 
 
 
- $user_sql1 = "DELETE FROM dispatch where dispatch.id = $dispatchid and dispatch.dispatch_noteid=$dispatch_noteid";
-   //$sql = "select * from login";
-   if ($conn->query($user_sql1)===TRUE) {
+      if ($canDelete>0) {
 
-          $user_sql2 = "update dispatch_note_total_dispatched set quantity=quantity-1 where dispatch_noteid=$dispatch_noteid";
+              $user_sql1 = "DELETE FROM dispatch where dispatch.id = $dispatchid and dispatch.dispatch_noteid=$dispatch_noteid";
          //$sql = "select * from login";
-         if ($conn->query($user_sql2)===TRUE) {
+         if ($conn->query($user_sql1)===TRUE) {
 
-          $temp=array("response"=>"success");
-          array_push($data1,$temp);
-           
+                $user_sql2 = "update dispatch_note_total_dispatched set quantity=quantity-1 where dispatch_noteid=$dispatch_noteid";
+               //$sql = "select * from login";
+               if ($conn->query($user_sql2)===TRUE) {
+
+                $temp=array("response"=>"success");
+                array_push($data1,$temp);
+                 
+                }
+
+              }
+       
+          }else{
+
+                $temp=array("response"=>"Cant Delete Bale From Dispatch Note");
+                array_push($data1,$temp);
+
           }
 
-        }
+
+      }
+
+
+
  
-    }
-
-
 
 
 echo json_encode($data1);
